Update README.md
Fantasy Delicacy is a unique food ordering platform that combines delicious dining with fantasy-inspired themes. Featuring an imaginative menu and user-friendly interface, Fantasy Delicacy lets users explore and order dishes that bring a touch of magic to every meal. The site is fully responsive and optimized for seamless browsing on any device.